pyproject.toml
[tool.poetry]
name = "e-learning platform"
version = "0.1.0"
description = ""
authors = ["seela-2001 <sohilamustafa68@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
pillow = "^11.0.0"
django-braces = "^1.16.0"
django-embed-video = "^1.4.10"
pymemcache = "^4.0.0"
django-debug-toolbar = "^4.4.6"
redis = "^5.2.0"
django-redisboard = "^8.4.0"
attrs = "^24.2.0"
djangorestframework = "^3.15.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
